backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
groq==0.13.1
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
numpy==2.2.1
sentence-transformers==3.3.1
pytesseract==0.3.10
pdf2image==1.17.0
PyJWT[crypto]==2.10.1
httpx==0.28.1
